feat: add function call tracing
The evaluator now traces function calls. Each call results in two trace messages
of the following formats:
smol1> FORM
smol1> FORM-WITH-EVALUATED-ARGUMENTS => VALUE

Here, FORM is before any sub-expressions have been evaluated, VALUE is the
result of the call, and FORM-WITH-EVALUATED-ARGUMENTS shows the operator along
with the evaluated arguments that were passed in.